# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Books Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get an only book",
                "description": "Get an only book",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of book to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "books"
                ],
                "summary": "Update a book",
                "description": "Update a book",
                "operationId": "f5ca658c501a803d68655af0365e3802",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of book to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "A name",
                                "value": "The theory of everything"
                            }
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Author of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "An author name",
                                "value": "Robin Sharma"
                            }
                        }
                    },
                    {
                        "name": "publisher",
                        "in": "query",
                        "description": "Publisher of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "A publisher company",
                                "value": "Jaiko Publishing House"
                            }
                        }
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Genre of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "A genre",
                                "value": "Fiction"
                            }
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Price of book",
                        "required": true,
                        "schema": {
                            "type": "int"
                        },
                        "examples": {
                            "int": {
                                "summary": "A price of book",
                                "value": "141"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "books"
                ],
                "summary": "Delete a book",
                "description": "Delete a book",
                "operationId": "1a9c26667c32b8ed5e5770d7a1cc9d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of book to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/books/": {
            "post": {
                "tags": [
                    "books"
                ],
                "summary": "Store a book",
                "description": "Store a book",
                "operationId": "9bf5f90833b3847b6168167f5ae5ced2",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "A name",
                                "value": "The theory of everything"
                            }
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Author of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "An author name",
                                "value": "Robin Sharma"
                            }
                        }
                    },
                    {
                        "name": "publisher",
                        "in": "query",
                        "description": "Publisher of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "A publisher company",
                                "value": "Jaiko Publishing House"
                            }
                        }
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Genre of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "A genre",
                                "value": "Fiction"
                            }
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Price of book",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "A price of book",
                                "value": "141"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "List all books",
                "description": "List all books",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    }
}